backend/pyproject.toml
[virtualenvs]
in-project = true

[tool.poetry]
name = "backend"
version = "0.1.0"
description = "AudioScribe backend service"
authors = ["vet-an@yandex.ru"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.3"}
psycopg2-binary = "^2.9.10"
redis = "^6.2.0"
celery = "^5.5.3"
loguru = "^0.7.3"
sqlalchemy = "^2.0.41"
pydantic-settings = "^2.9.1"
email-validator = "^2.2.0"
asyncpg = "^0.30.0"
alembic = "^1.16.2"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]
skip-string-normalization = false
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["Audioscribe_AL"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_equality = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
  "django.*",
  "celery.*",
]
ignore_missing_imports = true

[tool.ruff]
ignore = ["RUF001", "RUF002", "RUF003"]
